Use long double instead of float

+++ FFTBuffer.cpp
@@ -5,8 +5,8 @@ FFTBuffer::FFTBuffer(const int NFFT)
 {
     for (int i = 0; i < NFFT; ++i)
     {
-        data_[i].r = 0.0f;
-        data_[i].i = 0.0f;
+        data_[i].r = 0.0L;
+        data_[i].i = 0.0L;
     }
 }
 
@@ -15,8 +15,8 @@ FFTBuffer::FFTBuffer(const QByteArray& b)
 {
     for (int i = 0; i < b.size(); ++i)
     {
-        data_[i].r = static_cast<float>(b[i]);
-        data_[i].i = 0.0f;
+        data_[i].r = static_cast<long double>(b[i]);
+        data_[i].i = 0.0L;
     }
 }
 
@@ -25,8 +25,8 @@ FFTBuffer::FFTBuffer(const int* data, const int size)
 {
     for(int i = 0; i < size; ++i)
     {
-        data_[i].r = static_cast<float>(data[i]);
-        data_[i].i = 0.0f;
+        data_[i].r = static_cast<long double>(data[i]);
+        data_[i].i = 0.0L;
     }
 }
